Renames prefixed-db to prefixedDatabase
